Separated forgot password page components into separate files to solve rendering issue, aligned eye icons
