+++ package.json
@@ -20,6 +20,7 @@
       "axios": "^0.21.1",
       "bootstrap": "^4.6.0",
       "font-awesome": "^4.7.0",
+      "prop-types": "^15.7.2",
       "react": "^17.0.1",
       "react-bootstrap": "^1.4.3",
       "react-dom": "^17.0.1",
